[F] Custom derive for events (#26)
* Move to Cargo workspace

Procedural macros must be the only export from a crate, so the solution
is to make a separate crate with _just_ the proc macro exported.

* Add base work for custom derive

* Fix some missing imports

* Implement serialize for EventData structs

* Cleanup

* Event identifiers for serialized struct

* Deserialize enums with lots of caveats

* Fix quoting

* Fix test

* Squelch some warnings

* Squelch warnings, impl enum serialize

* Split tests up

* Implement some TODOs, add more tests

* "Fix" doc tests by removing them

* Get event type/ns for an event in an enum

* Remove incorrect TODO

If the event is in an enum, its types will be defined by that enum. If
it's just in a struct on its own, it may have different values. This is
by design.

* Refactor with Events and EventData struct

Structs have static methods that return their identifiers as strings.
They are implemented by the Events derive, so only structs in an
Events-derived enum will have EventData implemented on them.

* Squelch warnings

* Allow failures in nightly builds

* Mark struct derive as unused

* Cleanup

* Fix/add benchmarks

* Optimise event deserialize

* Remove double-deserialize
* Remove some clone()s

Speedup between 6 and 24%:

deserialize old style event
                        time:   [990.61 ns 1.0082 us 1.0264 us]
                        change: [-7.0474% -6.0965% -5.0652%] (p = 0.00 < 0.05)
                        Performance has improved.

deserialize event       time:   [822.15 ns 823.02 ns 824.02 ns]
                        change: [-27.182% -24.452% -22.543%] (p = 0.00 < 0.05)
                        Performance has improved.

* Improve deserialization code cleanliness

* Bit more cleanup

* Filter out own attributes

* Squelch warning

* Better tests

* Support rename attribute for enum variants

* Add serialize roundtrip benchmark

* Fmt

* Remove reliance on trait being in scope

Speeds serialization up _just slightly_ by using &str instead of String