[S] CMP-487 Clean DB tables before every PG test (#46)
* [S] Clean DB tables before every PG test

This PR introduces some macros to aid testing with stateful PG backing
stores. The idea is that events related to a particular test can be
removed from the database before the test is run (again) by their
`ident` field. It's not perfect, but as long as the `ident` used in each
test is different, there shouldn't be (m)any conflicts.

* Fix tests