Add event handler function
This is different to previous ways of handling events. Now, the event
handler is implemented as a trait method on the event to handle as
opposed to a callback being defined. This solves some lifetime/`move`
problems, making life a bit easier. It also binds the event handling _to
the event struct itself_ instead of some random closure, so this adds a
bit more type safety also.