[S] Unifier take 2 (#79)
* Update conflicting inserts

* Allow subject to be null

This mirrors the behaviour of the event store EventContext struct. An
empty subject should be null instead of `{}`.

* Fix duplicate key update syntax

* Map empty subject to '{}' in DB

For backwards compat reasons, '{}' should deserialize to None. Likewise,
None should serialize to '{}'. This should really be stored as null
instead of an empty object, but the decision was made a long time ago
and there is a lot of Typescript out there that depends on subject being
at least an object.

* Add ser/de tests